app/requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
